kv: deflake and unskip TestSecondaryTenantFollowerReadsRouting
Fixes #95338.

The test was failing to wait for configuration changes that added VOTER
replicas to apply on all replicas, causing flakiness. This commit fixes
that issue.

Before the fix, the test failed after 100 iterations on average on a
gceworker. After the fix, the test has not failed in over 2000 iterations.

Release note: None