server: ignore spanconfig limit for shared-process tenants
This change installs a noop spanconfig limiter when starting
shared-process tenants.  Since shared-process tenants are currently
used for "unified architecture" in which the tenant is expected to
behave as close as possible to the system tenant.

When we have a tenant-side capabilities reader, we should tie this to
a capability instead.

This has an unfortunate side-effect of making any tenant-level
override for the span limit a lie for shared-process tenants.

Fixes #93562

Release note: None