opt: fix internal error when exploration uncovers null-rejection
It is possible for exploration rules to reveal opportunities to
null-reject (simplify) an outer join that were not visible during
normalization. Previously, this would cause an internal error, since
it was assumed null-rejection would never be requested by an expression
that can prove the requested column is non-null. This patch removes the
assertion and instead removes any proven non-null columns from the
requested null-rejection column set. Note that this solution is
orthogonal to improving null-rejection during normalization - one
instance of this is tracked in #100564.

Fixes #100559

Release note (bug fix): Fixed a bug existing since before 22.1 that
could cause an internal error in rare cases for a query with outer joins
that can be simplified to non-outer joins and at least one semi-join.